NamCao - TEST - Cypress Setup & Basic Cypress Tests for Login and Registration
This commit sets up Cypress with the project. It also adds some basic examples of Cypress tests for inputs on the Login page and the Account Registration page, but much more will have to be done later.
To use Cypress testing, first run the CMS with 'node app.js'. Then, open a new terminal and run 'npm run cypress'. It will open up the Cypress application. Click on a spec file to run its tests.